+++ tests/dummy/config/environment.js
@@ -16,6 +16,12 @@ module.exports = function(environment) {
     APP: {
       // Here you can pass flags/options to your application instance
       // when it is created
+    },
+    contentSecurityPolicy: {
+      // cedar uses inline styles
+      'style-src': "'self' 'unsafe-inline'",
+      // connect to hosted arcigs servers
+      'connect-src': '*'
     }
   };