+++ CHANGELOG.md
@@ -8,8 +8,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+- update to use latest version of Cedar
+- removes deprecated (and potentially unsafe) binding to cedar-chart style attribute
 
 ### Support
+- added content security policies in dummy app b/c Cedar uses inline styles and remote services
 
 ## v0.2.2
 
@@ -32,7 +35,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - chart component listens for changes to new `override` property and re-draws chart
 
 ### Changed
-- removed support for `options.elementId` and getting id from 
+- removed support for `options.elementId` and getting id from
 
 ### Support
 - added route to show chart overrides